requirements.txt
# Flask лендинг — Московское Агентство Недвижимости
Flask>=3.0.0
requests>=2.28.0
gunicorn>=20.1.0
PyJWT>=2.8.0